Vols156-310 1st-pass descripts added; need cleanup
Added code to list to extract descriptions from -CATALOG; however, the original -CATALOGs are a mess of inconsistency, so it's a struggle to automate extraction. After vols 150 the formatting of -CATALOG settles down to more consistency, allowing an approximately clean stab at extraction. This has been used to add first-pass extracted descriptions to -CATALOG.ALL.